Rely on node 6's built-in ES6 support rather than using Babel.
This makes it much easier to debug, rather than messing about with source maps and the dearth of tooling that supports them. We can restore Babel later if we want to run within Firefox.

As a bonus, we don't need a build process anymore (for now).

Note that V8 (and thus Node) doesn't yet support the ES6 import syntax.