Apply the MPL.
Of the 3 licenses Mozilla pre-approves (MIT, MPL, Apache 2), MPL is the most restrictive, giving us the option of loosening it later. Its main stipulation is that modifications to the fathom.js file must be open-sourced, which I don't think is unreasonable. Let's see if it discourages too many people from using Fathom.