Move away from `var`.
My initial impression is this makes it easier to reason about the code. Block scope certainly decreases the amount of state you have to think about at any one time. There's no sense requiring functions or classes to be defined in specific orders, though; this isn't C, and they're still hoisted in ES6.

And my favorite: we no longer need to do a bunch of declaration bookkeeping at the front of each function just to accurately reflect the language semantics. It's less work and terser. Accurate reflection of the semantics is ensured by no-use-before-define.