Decide and document what is done with notes and scores when multiple LHSs emit the same node.
Different dom() selectors can overlap, leading to repeated applications of types, notes, and even score multipliers:

```
rule(dom('p'), type('paragraphish').score(10)),
     dom('p.intro'), type('paragraphish').score(5))
```

But actually, applying multiple score multipliers is fine. Maybe they want to give all <p>s a base score and add a special bonus for p.intros. The type() in the 2nd rule is necessary because dom() LHSs require a type.

We'll have meaningful notes override undefined or absent ones (as is currently done in InwardRule.results(). When you query for fnodes of a certain type, you can expect to find notes of any of form you specified on any rule with that type in the RHS.