Make RHS realize that calling func() means its emitted type is uninferrable.
Thus, raise an error when func() is used without typeIn().

Fixing the 'the fnode corresponding to a passed-in node' test to conform to obey this new rule uncovers a new bug we'll have to fix: typeIn() is checked too early, before we've taken into account the possibility of the type being inferred from a LHS in the case that func() does not emit a type.