Change scoreUpTo() to atMost().
This is the result of a lot of mental wrestling and discussion with potential users. Basically, I wanted to be more concise without being misleadingly inconsistent. Having max(4) meaning one thing ("top 4") on the LHS and another ("at most, the score 4") on the RHS was unacceptable. I'd actually prefer "top" for the LHS, but "max" is already in place in code and minds, and it's 90% as good. Other contenders for "atMost" were most() (whose inverse, least(), on the LHS would falsely suggest "the lowest ones") and under() (which is technically incorrect and so leads to having to read it and then readjust your thought each time). The LHS threshold operator will thus be atLeast().