Now you can ask any fnode for a score, note, or type, and you'll get the truth, even if it didn't happen to be computed yet.
This removes the user's need to care about Fathom's laziness, as far as I can recall right now.

As a bonus, this allows us to avoid eager-computing the entire ruleset when a ruleset is asked about a specific DOM node: now it can just happily return fnodes which compute on demand.

Factor out the non-eager bits into dedicated "soFar" routines on Fnode so tests that test our laziness could still call them and so, in one place, the executor wouldn't get into an infinite recursion.

Remove the TODO about enumerating the types on a fnode. We won't provide a way to enumerate them—at least not without a compelling demonstration that it's useful—as that would require running an entire ruleset to exhaustion. Making callers ask about types individually using hasType() lets us remain lazy. Making hasType() eager-compute the asked-about type makes this pattern usable for situations where you're interested in whether the page as a whole is classified as various types.