Confirm that passing in arbitrary LHSs to get() work in the new toposort-based executor.
The A->A rules trigger because the autocreated rule is always either A.max->* or A->*, both of which call in anything emitting A as a prereq.

Change the thing.asLhs condition just for consistency with the previous condition. It doesn't really matter, but I wasted time wondering why they were different.